requirements.txt
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.4
django-filter==2.2.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-simple-history==2.8.0
djangorestframework==3.10.2
gunicorn==20.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
xlwt==1.3.0

